Api/V1/PostCommentControllerにapi authorizationを実装 212
art make:policy CommentPolicy --model=Comment

+++ laravel/app/Http/Controllers/Api/V1/PostCommentController.php
@@ -69,6 +69,7 @@ class PostCommentController extends Controller
      */
     public function update(BlogPost $post, Comment $comment, StoreComment $request)
     {
+        $this->authorize($comment);
         $comment->content = $request->input('content');
         $comment->save();
         return new CommentResource($comment);
@@ -82,6 +83,7 @@ class PostCommentController extends Controller
      */
     public function destroy(BlogPost $post, Comment $comment)
     {
+        $this->authorize($comment);
         $comment->delete();
         // noContent methodでHTTPステータスコードの204 No Contentを返す
         return response()->noContent();
